Countdown: customizable snooze on the finish alarm
UP button quick-snoozes for the configured SnoozeSec length (label '+N Min'/'+1 h',
hidden + inert when Off, in which case Back = Stop). SELECT opens an ActionMenu of
1/3/5/10/15/30/45/60 min; picking one snoozes that timer and exits to the watchface.
Snooze length loaded from persist (key 8, default 60s) and updated from the inbox via
the type-tolerant parser. tc_format_snooze casts int32_t for %d (arm-none-eabi long).

+++ src/c/timer_calc.c
@@ -69,9 +69,9 @@ void tc_format_snooze(char *buf, size_t n, int32_t secs, bool with_plus) {
   const char *plus = with_plus ? "+" : "";
   const char *minu = with_plus ? "Min" : "min";
   if (secs >= 3600 && secs % 3600 == 0) {
-    snprintf(buf, n, "%s%d h", plus, secs / 3600);
+    snprintf(buf, n, "%s%d h", plus, (int)(secs / 3600));
   } else {
-    snprintf(buf, n, "%s%d %s", plus, secs / 60, minu);
+    snprintf(buf, n, "%s%d %s", plus, (int)(secs / 60), minu);
   }
 }
